Give plural named modules singular names for consistency
