+++ src/dependency_graph.rs
