+++ src/paths.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 
 use crate::config::Config;
-use crate::errors::{CargoMSRVError, IoErrorSource, TResult};
+use crate::error::{CargoMSRVError, IoErrorSource, TResult};
 
 pub fn crate_root_folder(config: &Config) -> TResult<PathBuf> {
     if let Some(path) = config.crate_path() {
