+++ src/search_method/linear.rs
@@ -6,7 +6,7 @@ use crate::{Config, TResult};
 // use crate::reporter::{write_failed_check, write_succeeded_check};
 use crate::reporter::Reporter;
 use crate::result::MinimalCompatibility;
-use crate::search_methods::FindMinimalCapableToolchain;
+use crate::search_method::FindMinimalCapableToolchain;
 use crate::toolchain::{OwnedToolchainSpec, ToolchainSpec};
 
 pub struct Linear<'runner, R: Check> {
