+++ src/sub_command/find.rs
@@ -2,15 +2,15 @@ use rust_releases::{Release, ReleaseIndex};
 
 use crate::check::Check;
 use crate::config::{Config, SearchMethod};
-use crate::errors::{CargoMSRVError, TResult};
+use crate::error::{CargoMSRVError, TResult};
+use crate::filter_releases::filter_releases;
 use crate::manifest::bare_version::BareVersion;
-use crate::releases::filter_releases;
 use crate::reporter::event::MsrvResult;
 use crate::reporter::Reporter;
 use crate::result::MinimalCompatibility;
-use crate::search_methods::{Bisect, FindMinimalCapableToolchain, Linear};
-use crate::writers::toolchain_file::write_toolchain_file;
-use crate::writers::write_msrv::write_msrv;
+use crate::search_method::{Bisect, FindMinimalCapableToolchain, Linear};
+use crate::writer::toolchain_file::write_toolchain_file;
+use crate::writer::write_msrv::write_msrv;
 use crate::{semver, SubCommand};
 
 pub struct Find<'index, C: Check> {
